Ensure that offsets are converted correctly.
The most important part here is the .0 at the last 1024. Without it, this was an integer division that was then converted to float, losing precision in the process. I've now added casts to all of these things, to ensure that this doesn't happen again and that I know for a fact that I'm doing a conversion here when looking at it.

+++ TR Poser/Rendering/TRRenderMoveableSCN.m
@@ -88,7 +88,7 @@
 	
 	__block NSUInteger angleSet = 0;
 	
-	self.rootNode.offset = SCNVector3Make(frame.offsetX / 1024.0, frame.offsetY / 1024.0, frame.offsetZ / 1024);
+	self.rootNode.offset = SCNVector3Make((CGFloat) frame.offsetX / 1024.0, (CGFloat) frame.offsetY / 1024.0, (CGFloat) frame.offsetZ / 1024.0);
 	
 	[self.rootNode enumerate:^(TRRenderMoveableNode *node){
 		node.rotationX = [frame rotationXInRadAtIndex:angleSet];
